pr comment: fix pub rand file
