+++ contracts/finality/src/exec/mod.rs
@@ -1,2 +1,3 @@
 pub mod admin;
 pub mod finality;
+pub mod public_randomness;
